models/gpt_guidance.py
"""
GPT-based guidance module for SurgicalAI.

This module provides GPT-based guidance for surgical procedures.
It requires the transformers package to be installed.
"""

import os
import json
import logging
import torch
import numpy as np
from typing import Dict, List, Optional, Any, Union
import time
import random

# Configure logging
logger = logging.getLogger(__name__)

# Handle missing transformers package
try:
    from transformers import (
        GPT2LMHeadModel,
        GPT2Tokenizer,
        GPT2Config,
        pipeline,
        AutoModelForCausalLM,
        AutoTokenizer
    )
    TRANSFORMERS_AVAILABLE = True
except ImportError:
    logger.warning("transformers package not installed. GPT-based models will not be available.")
    TRANSFORMERS_AVAILABLE = False
    # Create placeholder classes
    class GPT2LMHeadModel:
        def __init__(self, *args, **kwargs):
            raise ImportError("transformers package not installed")
    
    class GPT2Tokenizer:
        def __init__(self, *args, **kwargs):
            raise ImportError("transformers package not installed")
    
    class GPT2Config:
        def __init__(self, *args, **kwargs):
            raise ImportError("transformers package not installed")
    
    class AutoModelForCausalLM:
        @classmethod
        def from_pretrained(cls, *args, **kwargs):
            raise ImportError("transformers package not installed")
    
    class AutoTokenizer:
        @classmethod
        def from_pretrained(cls, *args, **kwargs):
            raise ImportError("transformers package not installed")
    
    def pipeline(*args, **kwargs):
        raise ImportError("transformers package not installed")


class GPTGuidance:
    """
    GPT-based guidance model for surgical procedures.
    
    This class provides personalized guidance for surgeons during procedures,
    leveraging pretrained GPT models fine-tuned on surgical data.
    """
    
    def __init__(self, 
                 model_name="gpt2",
                 weights_path=None,
                 procedure_knowledge_path=None,
                 max_length=128,
                 temperature=0.7,
                 device=None):
        """
        Initialize the GPT guidance model.
        
        Args:
            model_name: Name of pretrained model or path to model
            weights_path: Path to fine-tuned weights (optional)
            procedure_knowledge_path: Path to procedure knowledge data
            max_length: Maximum length of generated text
            temperature: Temperature for text generation
            device: Device to run the model on
        """
        if not TRANSFORMERS_AVAILABLE:
            logger.error("Cannot initialize GPT guidance - transformers package not installed")
            return
            
        self.model_name = model_name
        self.max_length = max_length
        self.temperature = temperature
        self.device = device if device is not None else torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        
        # Load procedure knowledge
        self.procedure_knowledge = {}
        if procedure_knowledge_path and os.path.exists(procedure_knowledge_path):
            try:
                with open(procedure_knowledge_path, 'r') as f:
                    self.procedure_knowledge = json.load(f)
                logger.info(f"Loaded procedure knowledge from {procedure_knowledge_path}")
            except Exception as e:
                logger.error(f"Failed to load procedure knowledge: {str(e)}")
                
        # Initialize fallback responses
        self._init_fallback_responses()
        
        # Load model and tokenizer
        try:
            if weights_path and os.path.exists(weights_path):
                # Load fine-tuned model
                logger.info(f"Loading fine-tuned GPT model from {weights_path}")
                self.model = GPT2LMHeadModel.from_pretrained(weights_path)
                self.tokenizer = GPT2Tokenizer.from_pretrained(weights_path)
            else:
                # Load pretrained model
                logger.info(f"Loading pretrained GPT model: {model_name}")
                self.model = GPT2LMHeadModel.from_pretrained(model_name)
                self.tokenizer = GPT2Tokenizer.from_pretrained(model_name)
                
            self.model.to(self.device)
            self.model.eval()
            
            # Set up generation pipeline
            self.generator = pipeline(
                "text-generation",
                model=self.model,
                tokenizer=self.tokenizer,
                device=0 if str(self.device) == "cuda" else -1
            )
            
            logger.info("GPT guidance model initialized successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize GPT model: {str(e)}")
            self.model = None
            self.tokenizer = None
            self.generator = None

    def _init_fallback_responses(self):
        # Implementation of _init_fallback_responses method
        pass

    def format_prompt(self, 
                     phase: str, 
                     tools: List[str], 
                     anatomical_structures: List[str],
                     cvs_achieved: bool,
                     previous_actions: List[str],
                     experience_level: str = "intermediate",
                     common_mistakes: List[str] = None) -> str:
        """
        Format the prompt for guidance generation.
        
        Args:
            phase: Current surgical phase
            tools: List of detected tools
            anatomical_structures: List of visible anatomical structures
            cvs_achieved: Whether Critical View of Safety is achieved
            previous_actions: List of previous actions
            experience_level: Surgeon experience level
            common_mistakes: List of common mistakes in this phase
            
        Returns:
            Formatted prompt string
        """
        # Implementation of format_prompt method
        pass

    def generate_guidance(self, 
                          prompt: str, 
                          max_length: int = 100,
                          temperature: float = 0.7,
                          top_p: float = 0.9,
                          top_k: int = 50,
                          num_return_sequences: int = 1) -> List[str]:
        """
        Generate guidance based on the prompt.
        
        Args:
            prompt: Formatted prompt
            max_length: Maximum length of the generated text
            temperature: Temperature for sampling
            top_p: Top-p sampling parameter
            top_k: Top-k sampling parameter
            num_return_sequences: Number of guidance options to generate
            
        Returns:
            List of generated guidance texts
        """
        # Implementation of generate_guidance method
        pass

    def get_personalized_guidance(self, 
                                  context: Dict[str, Any],
                                  user_profile: Any = None) -> Dict[str, Any]:
        """
        Get personalized guidance based on the current context and user profile.
        
        Args:
            context: Dictionary with surgical context
            user_profile: User profile object (optional)
            
        Returns:
            Dictionary with guidance information
        """
        # Implementation of get_personalized_guidance method
        pass

    def get_phase_guidance(self, 
                          phase_name: str, 
                          is_transition: bool = True,
                          user_profile: Any = None) -> str:
        """
        Get guidance for a specific surgical phase.
        
        Args:
            phase_name: Name of the surgical phase
            is_transition: Whether this is a transition to the phase
            user_profile: User profile for personalization
            
        Returns:
            Phase guidance text
        """
        # Implementation of get_phase_guidance method
        pass

    def get_mistake_guidance(self,
                            mistake_info: Dict[str, Any],
                            context: Dict[str, Any],
                            user_profile: Any = None) -> str:
        """
        Get guidance for handling a detected mistake.
        
        Args:
            mistake_info: Information about the detected mistake
            context: Current surgical context
            user_profile: User profile for personalization
            
        Returns:
            Mistake guidance text
        """
        # Implementation of get_mistake_guidance method
        pass

    def get_tool_guidance(self, 
                         current_tools: List[str],
                         recommended_tools: List[str],
                         phase_name: str,
                         user_profile: Any = None) -> str:
        """
        Get guidance for tool usage.
        
        Args:
            current_tools: Currently detected tools
            recommended_tools: Recommended tools for this phase
            phase_name: Current surgical phase
            user_profile: User profile for personalization
            
        Returns:
            Tool guidance text
        """
        # Implementation of get_tool_guidance method
        pass

    def save(self, path: str) -> bool:
        """
        Save the model and tokenizer.
        
        Args:
            path: Directory path to save to
            
        Returns:
            True if save was successful
        """
        # Implementation of save method
        pass

    def load(self, path: str) -> bool:
        """
        Load the model and tokenizer.
        
        Args:
            path: Directory path to load from
            
        Returns:
            True if load was successful
        """
        # Implementation of load method
        pass 