requirements.txt
# Core dependencies for SurgicalAI training
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
PyYAML>=6.0
tqdm>=4.62.0

# PyTorch ecosystem
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.12

# Computer vision
opencv-python-headless>=4.5.0
albumentations>=1.1.0
Pillow>=9.0.0

# Data processing
pycocotools>=2.0.4

# Training utilities
tensorboard>=2.7.0
einops>=0.4.1
torchmetrics>=0.8.0 